pyproject.toml
[project]
name = "autoinsta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asttokens>=3.0.0",
    "beautifulsoup4>=4.13.3",
    "comm>=0.2.2",
    "fastapi[standard]>=0.115.8",
    "lxml>=5.3.1",
    "pycryptodome>=3.21.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
